+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## V4.1.6
 
-#### Date: TBD
+#### Date: 3/6/2026
 
 ### :rocket: Features