+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 -   Added the `--aux-version <version>` option to the `pack-aux` CLI command.
 -   Added the [`@preact/signals`](https://www.npmjs.com/package/@preact/signals) library.
     -   This also includes `@preact/signals/utils`.
+-   Added `preact/hooks` as an importable library.
 
 ### :bug: Bug Fixes